✨ Implement `ambient_color`

+++ src/args.rs
@@ -1,6 +1,6 @@
 use std::path::PathBuf;
 
-use clap::Parser;
+use clap::{value_parser, Parser};
 
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about)]
@@ -21,6 +21,6 @@ pub struct Args {
 pub struct TracerOptions {
     /// Samples per pixel that get averaged for the antialiasing.
     /// When equals to `1`, no randomization for ray direction is applied.
-    #[arg(short, long, default_value = "1", value_parser = 1..)]
-    pub samples_per_pixel: usize,
+    #[arg(short, long, default_value = "1", value_parser = value_parser!(u16).range(1..))]
+    pub samples_per_pixel: u16,
 }
